pyproject.toml
[tool.poetry]
name = "venvai"
version = "0.1.0"
description = "AiCapilot"
authors = ["TheHormat"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.10.15"
django = "^5.1.3"
google-generativeai = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
